+++ pyfiles/MoveOutletsToStreams.py
@@ -64,6 +64,7 @@ for line in process.stdout.readlines():
 
 #arcpy.DefineProjection_management(om, coord_sys)
 # remove converted json file
-extn_json=os.path.splitext(shfl)[1] # get extension of the converted json file
-if extn_json==".json":
+if arcpy.Exists(inlyr2):
+  extn_json=os.path.splitext(shfl)[1] # get extension of the converted json file
+  if extn_json==".json":
     os.remove(shfl)
